Remove useFetch hook, use @tanstack/react-query and React's Suspense for fetching
Removed custom `useFetch` hook from the application codebase and replaced the `useFetch` instances with `useQuery` of `@tanstack/react-query`, which is a widely used library, robust, and provides better caching, refetching, and synchronization strategies.
Since `useQuery` supports suspense everywhere, including components that are using useFetch do not need to cater loading and error state themselves. Hence, initial state handling was also updated across for components using data fetching.
Also, removed the dependencies of `@ethang/fetch` from the certain components, replaced the `cacheBust` with query invalidation from `react-query`.

A custom `getRequestKey` helper was added to the `api-requests` module, which constructs a list of stringified request properties to be used as `react-query` unique keys of requests.

Updated the `zodFetch` method to use `fetch()` method.

The order of Providers were interchanged while rendering to prevent any redundancy.

Overall, this makes fetching data in the components more declarative and uniform, makes it utilize Suspense which will lead to better user experience during data fetching.

+++ src/app/(routes)/manage/(components)/delete-list-modal.tsx
@@ -1,14 +1,14 @@
 'use client';
 import { useUser } from '@clerk/nextjs';
-import { cacheBust } from '@ethang/fetch';
 import { useToggle } from '@ethang/hooks/use-toggle';
 import { useMutation } from '@tanstack/react-query';
 import type { JSX } from 'react';
 
 import { Button } from '../../../(components)/(elements)/button';
 import { Modal } from '../../../(components)/modal';
+import { queryClient } from '../../../(components)/providers';
 import { api } from '../../../data/api';
-import { apiRequests } from '../../../data/api-requests';
+import { apiRequests, getRequestKey } from '../../../data/api-requests';
 
 type DeleteListModalProperties = {
   readonly listId: string;
@@ -24,12 +24,18 @@ export function DeleteListModal({
 
   const { isLoading, mutate } = useMutation({
     async mutationFn() {
-      await api.deleteList(listId);
+      if (user?.id !== undefined) {
+        await api.deleteList(listId);
 
-      await Promise.all([
-        cacheBust(apiRequests.getManageLists(user?.id ?? '')),
-        cacheBust(apiRequests.getListPage()),
-      ]);
+        await Promise.all([
+          queryClient.invalidateQueries(
+            getRequestKey(apiRequests.getManageLists(user.id)),
+          ),
+          queryClient.invalidateQueries(
+            getRequestKey(apiRequests.getListPage()),
+          ),
+        ]);
+      }
 
       toggleOpen();
     },
